Fixing common responses.

+++ Get-DDnsUpdate/run.ps1
@@ -23,17 +23,17 @@ $auth = [System.Text.Encoding]::ASCII.GetString([System.Convert]::FromBase64Stri
 $username = $auth[0]
 $password = $auth[1]
 
-Write-Debug "Authorization: $username / ********" -Debug
+Write-Debug "Authorization: $username / ********"
 
 if (-not $env:AppUsername -or -not $env:AppPassword) {
     Write-Error "No credentials have been set, please ensure the username and password are set within the application settings."
 
     Push-OutputBinding -Name Response -Value ([HttpResponseContext]@{
-        StatusCode = [HttpStatusCode]::Unauthorized
+        StatusCode = [HttpStatusCode]::InternalServerError
     })
 
     exit
-} elseif (-not $username -eq $env:AppUsername -or -not $password -eq $env:AppPassword) {
+} elseif ($username -ne $env:AppUsername -or $password -ne $env:AppPassword) {
     Write-Error "The credentials did not match those configured in the application settings. Please check your configuration and try again."
 
     Push-OutputBinding -Name Response -Value ([HttpResponseContext]@{
@@ -59,13 +59,14 @@ if (-not $hostname) {
     Write-Error "The hostname was not provided in the request."
 
     Push-OutputBinding -Name Response -Value ([HttpResponseContext]@{
-        StatusCode = [HttpStatusCode]::BadRequest
+        StatusCode = [HttpStatusCode]::OK
+        Body = "nohost"
     })
 
     exit
 }
 
-Write-Debug "Hostname: $hostname" -Debug
+Write-Debug "Hostname: $hostname"
 
 $ipAddr = $Request.Query.MyIP
 if (-not $ipAddr) {
@@ -78,10 +79,10 @@ if (-not $ipAddr) {
     exit
 }
 
-Write-Debug "New IP Address: $ipAddr" -Debug
+Write-Debug "New IP Address: $ipAddr"
 
 $count = [regex]::matches($hostname, '[\.]').count
-Write-Debug "Found $count periods (.) within hostname '$hostname' provided." -Debug
+Write-Debug "Found $count periods (.) within hostname '$hostname' provided."
 
 # Determines whether a TLD has been provided using the number of periods included in the hostname.
 if ($count -eq 1) {
@@ -92,16 +93,17 @@ if ($count -eq 1) {
     $zoneName = $hostname.Substring($hostname.IndexOf('.') + 1)
 }
 
-Write-Debug "Name: $dnsName" -Debug
-Write-Debug "DNS Zone: $zoneName" -Debug
-Write-Debug "Resource Group: $dnsZoneRGName" -Debug
+Write-Debug "Name: $dnsName"
+Write-Debug "DNS Zone: $zoneName"
+Write-Debug "Resource Group: $dnsZoneRGName"
 
 $rs = Get-AzDnsRecordSet -ResourceGroupName $dnsZoneRGName -ZoneName $zoneName -Name $dnsName -RecordType A
 if (-not $rs) {
     Write-Error "Could not locate the DNS record '$dnsName' in zone '$zoneName'. Please check your Azure DNS configuration and try again."
 
     Push-OutputBinding -Name Response -Value ([HttpResponseContext]@{
-        StatusCode = [HttpStatusCode]::NotFound
+        StatusCode = [HttpStatusCode]::OK
+        Body = "nofqdn"
     })
 
     exit
@@ -110,44 +112,66 @@ if (-not $rs) {
     Write-Error "Could not update the DNS record '$dnsName' in zone '$zoneName' because it is an alias. Please check your Azure DNS configuration and try again."
 
     Push-OutputBinding -Name Response -Value ([HttpResponseContext]@{
-        StatusCode = [HttpStatusCode]::BadRequest
+        StatusCode = [HttpStatusCode]::OK
+        Body = "nofqdn"
     })
 
     exit
 }
 
-Write-Debug "Checking the existing records for zone '$zoneName'..." -Debug
+Write-Debug "Checking the existing records for zone '$zoneName'..."
 
 $found = $false
 $ipAddrsToRemove = @()
 
 foreach ($record in $rs.Records) {
     if ($record.Ipv4Address -ne $ipAddr) {
-        Write-Debug "Found IP address $record.Ipv4Address which does not belong in the record set..." -Debug
+        Write-Debug "Found IP address $record.Ipv4Address which does not belong in the record set..."
         $ipAddrsToRemove += $record.Ipv4Address
     } else {
-        Write-Debug "Expected IP address already exists within the record set..." -Debug
+        Write-Debug "Expected IP address already exists within the record set..."
         $found = $true
     }
 }
 
 Write-Information "Preparing to update the DNS zone '$zoneName' in Resource Group '$dnsZoneRGName'..."
+$saveChanges = $false
 
-foreach ($existingIpAddr in $ipAddrsToRemove) {
-    Remove-AzDnsRecordConfig -RecordSet $rs -Ipv4Address $existingIpAddr
-    Write-Information "Removed IPv4 address '$existingIpAddr' from DNS zone: '$zoneName'."
+if ($ipAddrsToRemove.count -gt 0) {
+    Write-Debug "Removing unnecessary IPv4 addresses..."
+
+    foreach ($existingIpAddr in $ipAddrsToRemove) {
+        Remove-AzDnsRecordConfig -RecordSet $rs -Ipv4Address $existingIpAddr
+        Write-Information "Removed IPv4 address '$existingIpAddr' from DNS zone: '$zoneName'."
+    }
+
+    $saveChanges = $true
 }
 
 if (!$found) {
     Add-AzDnsRecordConfig -RecordSet $rs -Ipv4Address $ipAddr
     Write-Information "Added new IPv4 Address '$ipAddr' to DNS zone '$zoneName'."
+
+    $saveChanges = $true
 }
 
-Set-AzDnsRecordSet -RecordSet $rs
-Write-Information "Successfully updated DNS zone '$zoneName' in Resource Group '$dnsZoneRGName'."
+if ($saveChanges) {
+    Set-AzDnsRecordSet -RecordSet $rs
+    Write-Information "Successfully updated DNS zone '$zoneName' in Resource Group '$dnsZoneRGName'."
+    
+    Push-OutputBinding -Name Response -Value ([HttpResponseContext]@{
+        StatusCode = [HttpStatusCode]::OK
+        Body       = "good"
+    })
 
-# The response values returned here are required by the Inadyn client, do not change!
-Push-OutputBinding -Name Response -Value ([HttpResponseContext]@{
-    StatusCode = [HttpStatusCode]::OK
-    Body       = "good"
-})
+    exit
+} else {
+    Write-Information "No change required for DNS zone '$zoneName' in Resource Group '$dnsZoneRGName'."
+
+    Push-OutputBinding -Name Response -Value ([HttpResponseContext]@{
+        StatusCode = [HttpStatusCode]::OK
+        Body       = "nochg"
+    })
+
+    exit
+}
